Extension of MCMCDiagnosticTools

The following methods of MCMCDiagnosticTools.jl are extended by this package.

MCMCDiagnosticTools.bfmi

MCMCDiagnosticTools.ess_rhat

MCMCDiagnosticTools.mcse

MCMCDiagnosticTools.rstar

MCMCDiagnosticTools.bfmi — Method

bfmi(data::InferenceData) -> DimArray
bfmi(sample_stats::Dataset) -> DimArray

Calculate the chainwise estimated Bayesian fraction of missing information (BFMI).

See MCMCDiagnosticTools.bfmi for more details.

MCMCDiagnosticTools.ess_rhat — Method

ess_rhat(data::InferenceData; kwargs...) -> Dataset
ess_rhat(data::Dataset; kwargs...) -> Dataset

Calculate the effective sample size (ESS) and $\widehat{R}$ diagnostic for each parameter in the data.

For more details and a description of the kwargs, see MCMCDiagnosticTools.ess_rhat.

MCMCDiagnosticTools.mcse — Method

mcse(data::InferenceData; kwargs...) -> Dataset
mcse(data::Dataset; kwargs...) -> Dataset

Calculate the Monte Carlo standard error (MCSE) for each parameter in the data.

For more details and a description of the kwargs, see MCMCDiagnosticTools.mcse.

MCMCDiagnosticTools.rstar — Method

rstar(
    rng::Random.AbstractRNG=Random.default_rng(),
    classifier,
    data::Union{InferenceData,Dataset};
    kwargs...,
)

Calculate the $R^*$ diagnostic for the data.

For a description of the classifier and kwargs, see MCMCDiagnosticTools.rstar.