Internal

PSIS.GeneralizedParetoKnownMu — Type

GeneralizedParetoKnownMu(μ)

Represents a GeneralizedPareto where $\mu$ is known.

PSIS.GeneralizedParetoKnownMuTheta — Type

GeneralizedParetoKnownMuTheta(μ, θ)

Represents a GeneralizedPareto where $\mu$ and $\theta=\frac{\xi}{\sigma}$ are known.

PSIS.plotting_backend! — Method

plotting_backend!(backend::Symbol)

Set default plotting backend. Valid values are :Plots and :Makie.

StatsBase.fit — Method

fit(g::GeneralizedParetoKnownMu, x; kwargs...)

Fit a GeneralizedPareto with known location μ to the data x.

The fit is performed using the Empirical Bayes method of [ZhangStephens2009][Zhang2010].

Keywords

sorted::Bool=issorted(x): If true, x is assumed to be sorted. If false, a sorted copy of x is made.

improved::Bool=true: If true, use the adaptive empirical prior of [Zhang2010]. If false, use the simpler prior of [ZhangStephens2009].

min_points::Int=30: The minimum number of quadrature points to use when estimating the posterior mean of $\theta = \frac{\xi}{\sigma}$.

ZhangStephens2009Jin Zhang & Michael A. Stephens (2009) A New and Efficient Estimation Method for the Generalized Pareto Distribution, Technometrics, 51:3, 316-325, DOI: 10.1198/tech.2009.08017

Zhang2010Jin Zhang (2010) Improving on Estimation for the Generalized Pareto Distribution, Technometrics, 52:3, 335-339, DOI: 10.1198/TECH.2010.09206