DimensionalData.jl

DimensionalData.jl

DimensionalData.DimensionalData — Module.

DimensionalData.jl provides types and methods for indexing with named dimensions, using named dimensions in Base and Statistics methods instead of Integer dims, and selecting data from dimension values instead of using indices directly.

Dimensions are simply types that wrap values. They both store dimension values and are used for dimension lookup or indices, ranges or dimension number. X, Y,ZandTime` are the unexported defaults, add this line to use them:

using DimensionalData: X, Y, Z, Time

Selectors find indices in the dimension based on values At, Near, or Between the passed in value(s).

These are some examples of valid syntax:

# Indexing with dim wrappers
a[X(1:10), Y(1:4)]
# Dim wrappers and a selector
a[X(1:10), Y<|At(25.7)]
# Unitful.jl selectors. Without dim wrappers selectors must be in the right order
a[Near(23"s"), Between(10.5u"m", 50.5u"m")]
# Dim type used instead of a dimension number
mean(a; dims=X)

source

DimensionalData.AbstractDimension — Type.

An AbstractDimension tags the dimensions in an AbstractArray.

It can also contain spatial coordinates and their metadata. For simplicity, the same types are used both for storing dimension information and for indexing.

source

DimensionalData.At — Type.

At(x)

Selector that exactly matches the value on the passed-in dimensions, or throws an error. For ranges and arrays, every value must match an existing value - not just the end points.

source

DimensionalData.Between — Type.

Between(a, b)

Selector that retreive all indices located between 2 values.

source

DimensionalData.Dim — Type.

A generic dimension. For use when custom dims are required when loading data from a file. The sintax is ugly and verbose to use for indexing, ie Dim{:lat}(1:9) rather than Lat(1:9). This is the main reason they are not the only type of dimension availabile.

source

DimensionalData.DimensionalArray — Type.

A basic DimensionalArray type

Maintains and updates its dimensions through transformations

source

DimensionalData.Near — Type.

Near(x)

Selector that selects the nearest index to its contained value(s)

source

DimensionalData.Selector — Type.

Selectors indicate that index values are not indices, but points to be selected from the dimension values, such as DateTime objects on a Time dimension.

source

DimensionalData.dims — Function.

dims(x)

Return a tuple of the dimensions for a dataset. These can contain the coordinate ranges if bounds() and select() are to be used, or you want them to be shown on plots in place of the array indices.

They can also contain a units string or unitful unit to use and plot dimension units.

This is the only method required for this package to work. It probably requires defining a dims field on your object to store dims in.

source

DimensionalData.dims — Method.

Get the dimension(s) matching the type(s) of the lookup dimension.

source

DimensionalData.label — Function.

label(x)

Get a plot label for data or a dimension. This will include the name and units if they exist, and anything else that should be shown on a plot.

source

DimensionalData.metadata — Function.

metadata(x)

Return the metadata of a dimension or data object.

source

DimensionalData.name — Function.

name(x)

Get the name of data or a dimension.

source

DimensionalData.refdims — Function.

refdims(x)

Reference dimensions for an array that is a slice or view of another array with more dimensions.

slicedims(a, dims) returns a tuple containing the current new dimensions and the new reference dimensions. Refdims can be stored in a field or disgarded, as it is mostly to give context to plots. Ignoring refdims will simply leave some captions empty.

source

DimensionalData.shortname — Function.

shortname(x)

Get the short name of array data or a dimension.

source

DimensionalData.units — Function.

units(x)

Return the units of a dimension. This could be a string, a unitful unit, or nothing.

source

DimensionalData.val — Function.

val(x)

Return the contained value of a wrapper object, otherwise just returns the object.

source

DimensionalData.AbstractDimCombination — Type.

AbstractCombined holds mapping that require multiple dimension when select() is used, shuch as for situations where they share an affine map or similar transformation instead of linear maps. Each dim that shares the mapping must contain the same (identical) object.

Dimensions holding a DimCombination will work as usual for direct indexing.

All AbstractDimension are assumed to val and metadata fields.

source

DimensionalData.AbstractParametricDimension — Type.

Dimensions with user-set type paremeters

source

DimensionalData.Forward — Type.

Trait indicating that the dimension is in the normal forward order.

source

DimensionalData.Reverse — Type.

Trait indicating that the dimension is in the reverse order. Selector lookup and plotting will be reverse.

source

Base.permutedims — Method.

Sort dimensions into the order they take in the array.

Missing dimensions are replaced with nothing

source

DimensionalData.dimnum — Method.

Get the number of an AbstractDimension as ordered in the array

source

DimensionalData.dims2indices — Function.

Convert a tuple of AbstractDimension to indices, ranges or Colon.

source

DimensionalData.formatdims — Method.

Format the dimension to match internal standards.

Mostily this means converting tuples and UnitRanges to LinRange, which is easier to handle.

Errors are thrown if dims don't match the array dims or size.

source

DimensionalData.rebuild — Function.

rebuild(x, data, [dims], [refdims])

Rebuild an object struct with an updated value

source

DimensionalData.reducedims — Method.

Replace the specified dimensions with an index of 1 to match a new array size where the dimension has been reduced to a length of 1, but the number of dimensions has not changed.

Used in mean, reduce, etc.

source

DimensionalData.slicedims — Method.

Slice the dimensions to match the axis values of the new array

All methods returns a tuple conatining two tuples: the new dimensions, and the reference dimensions. The ref dimensions are no longer used in the new struct but are useful to give context to plots.

Called at the array level the returned tuple will also include the previous reference dims attached to the array.

source

DimensionalData.@dim — Macro.

@dim typ [name=string(typ)] [shortname=string(typ)]

Macro to easily define specific dimensions.

Example:

@dim Lat "Lattitude"

source